+++ lib/bloc/present_signed_document_cubit.dart
@@ -23,7 +23,7 @@ export 'present_signed_document_state.dart';
 /// which can be shared.
 @injectable
 class PresentSignedDocumentCubit extends Cubit<PresentSignedDocumentState> {
-  static final _log = Logger("PresentSignedDocumentCubit");
+  static final _log = Logger((PresentSignedDocumentCubit).toString());
   static final _tsDateFormat = DateFormat('yyyyMMddHHmmss');
 
   final AppService _appService;
@@ -38,7 +38,8 @@ class PresentSignedDocumentCubit extends Cubit<PresentSignedDocumentState> {
 
   /// Saves [signedDocument] into public directory.
   Future<void> saveDocument() async {
-    _log.info("Saving signed document: ${signedDocument.filename}.");
+    _log.info(
+        "Saving signed document: ${File(signedDocument.filename).redactedInfo}.");
 
     emit(state.toLoading());
 
@@ -50,12 +51,12 @@ class PresentSignedDocumentCubit extends Cubit<PresentSignedDocumentState> {
       // Need to change PresentSignedDocumentSuccessState impl. to allow File?
       await _saveDocumentIntoFile(file!);
 
-      _log.info("Signed Document was saved into $file");
+      _log.info("Signed Document was saved into ${file.redactedInfo}");
 
       emit(state.toSuccess(file));
     } catch (error, stackTrace) {
-      _log.severe(
-          "Error saving signed Document into $file.", error, stackTrace);
+      _log.severe("Error saving signed Document into ${file?.redactedInfo}.",
+          error, stackTrace);
 
       emit(state.toError(error));
     }