+++ lib/ui/screens/main_screen.dart
@@ -14,6 +14,7 @@ import '../../app_service.dart';
 import '../../bloc/app_bloc.dart';
 import '../../deep_links.dart';
 import '../../di.dart';
+import '../../file_extensions.dart';
 import '../../services/encryption_key_registry.dart';
 import '../../strings_context.dart';
 import '../app_theme.dart';
@@ -126,6 +127,9 @@ class _MainScreenState extends State<MainScreen> {
 
     return showGeneralDialog(
       context: context,
+      routeSettings: RouteSettings(
+        name: screen.runtimeType.toString(),
+      ),
       pageBuilder: (context, _, __) => screen,
     );
   }
@@ -205,9 +209,9 @@ class _MainScreenState extends State<MainScreen> {
     final selectedFile = result?.files.singleOrNull;
 
     if (selectedFile != null) {
-      final File file = File(selectedFile.path!);
+      final file = File(selectedFile.path!);
 
-      _logger.fine('File selected: $file');
+      _logger.fine('File selected: ${file.redactedInfo}');
 
       if (context.mounted) {
         _openNewFile(file);