+++ lib/utils.dart
@@ -3,6 +3,7 @@ import 'dart:math' show Random;
 
 import 'package:basic_utils/basic_utils.dart'
     show X509CertificateData, X509Utils;
+import 'package:logging/logging.dart' show LogRecord, Level;
 
 final Random _random = Random.secure();
 
@@ -25,4 +26,20 @@ X509CertificateData x509CertificateDataFromDer(String data) {
   ].join("\n");
 
   return X509Utils.x509CertificateFromPem(pem);
+}
+
+/// Format given [log] as single message.
+String formatCrashlyticsLog(final LogRecord log) {
+  final level = switch (log.level) {
+    Level.FINEST || Level.FINER || Level.FINE || Level.CONFIG => 'D',
+    Level.INFO => 'I',
+    Level.WARNING => 'W',
+    Level.SEVERE || Level.SHOUT => 'E',
+    _ => "?"
+  };
+  final tag = log.loggerName;
+  final line1 = "${log.time}: $level/$tag: ${log.message}";
+  final error = log.error;
+
+  return (error != null ? "$line1\n$error" : line1);
 }